Change network name in upstream AN
This updates the commands for Observer to point to the currently latest AN8. 

However this will become outdated again the next time there is a network upgrade. It would be ideal if we set up a DNS for the observer upstream that persists across sporks, like we have for `access.mainnet.nodes.onflow`